pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "probe-ai"
version = "0.2.0"
description = "Chaos engineering for AI agents — generate failure scenarios, inject them into tool calls, and see what breaks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "IrisDataLabs"},
]
keywords = ["ai", "agents", "testing", "chaos-engineering", "evaluation", "resilience", "llm", "reliability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2.0", "langchain-core>=0.2.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.urls]
Homepage = "https://github.com/irisdatalabs/probe-ai"
Repository = "https://github.com/irisdatalabs/probe-ai"
Issues = "https://github.com/irisdatalabs/probe-ai/issues"

[project.scripts]
probe-ai = "agentprobe.cli:main"

[tool.setuptools.packages.find]
include = ["agentprobe*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
